+++ aide/src/main/kotlin/eu/vendeli/aide/AideCompilerPluginRegistrar.kt
@@ -1,0 +1,22 @@
+package eu.vendeli.aide
+
+import com.google.auto.service.AutoService
+import eu.vendeli.aide.ir.SendAutoAppenderExtension
+import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
+import org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
+import org.jetbrains.kotlin.compiler.plugin.ExperimentalCompilerApi
+import org.jetbrains.kotlin.config.CompilerConfiguration
+import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrarAdapter
+
+@OptIn(ExperimentalCompilerApi::class)
+@AutoService(CompilerPluginRegistrar::class)
+class AideCompilerPluginRegistrar : CompilerPluginRegistrar() {
+    override fun ExtensionStorage.registerExtensions(configuration: CompilerConfiguration) {
+        val doAutoSend = checkNotNull(configuration[AideArgumentCommandLineProcessor.AUTO_SEND_OPTION_KEY]).toBoolean()
+
+        FirExtensionRegistrarAdapter.registerExtension(AideFirExtensionRegistrar(doAutoSend))
+        IrGenerationExtension.registerExtension(SendAutoAppenderExtension(doAutoSend))
+    }
+
+    override val supportsK2: Boolean get() = true
+}