+++ api-sentinel/src/jvmMain/kotlin/eu/vendeli/sentinel/Utils.kt
@@ -1,4 +1,4 @@
-package eu.vendeli.ksp
+package eu.vendeli.sentinel
 
 import com.google.devtools.ksp.KspExperimental
 import com.google.devtools.ksp.processing.KSPLogger