+++ ktgram-gradle-plugin/src/main/kotlin/eu/vendeli/ktgram/gradle/KtGramExt.kt
@@ -15,6 +15,8 @@ import javax.inject.Inject
  * @property autoAnswerCallback set to true to automatically respond to callbacks unless otherwise specified in the annotation.
  * @property ktorJvmEngine the Ktor JVM engine to use, defaults to [KtorJvmEngine.JAVA].
  * @property handleLoggingProvider set to false to disable the logging provider handling.
+ * @property aideEnabled set to true to enable compilation Aide.
+ * @property aideAutoSend set to true to automatically send actions.
  */
 abstract class KtGramExt
     @Inject
@@ -28,4 +30,6 @@ abstract class KtGramExt
         val autoAnswerCallback = factory.property<Boolean>()
         val ktorJvmEngine = factory.property<KtorJvmEngine>()
         val handleLoggingProvider = factory.property<Boolean>()
+        val aideEnabled = factory.property<Boolean>()
+        val aideAutoSend = factory.property<Boolean>()
     }